simplify quality validator to meet loc limit
