+++ agent_workspaces/Agent-7/quality_validation.py
@@ -1,1462 +1,31 @@
 #!/usr/bin/env python3
-"""
-Quality Validation Script - Agent-7 (Quality Assurance Manager)
+"""Quality validation orchestrator."""
 
-Automated quality validation script that implements quality gates and validation
-checkpoints for cleanup operations. This script ensures all quality standards
-are met before, during, and after cleanup operations.
-
-Author: Agent-7 (Quality Assurance Manager)
-Task: Quality Systems Cleanup and Standards Establishment
-V2 Standards: ≤400 LOC, SRP, OOP principles
-"""
-
-import os
 import sys
-import time
-import subprocess
-import json
-import logging
-from pathlib import Path
-from typing import Dict, List, Any, Optional, Tuple
-from dataclasses import dataclass, field
-from enum import Enum
-import ast
-import inspect
-import re
 
-# Configure logging
-logging.basicConfig(
-    level=logging.INFO,
-    format='%(asctime)s [%(levelname)8s] %(name)s: %(message)s',
-    datefmt='%Y-%m-%d %H:%M:%S'
-)
-logger = logging.getLogger(__name__)
+from data_loader import get_project_root
+from reporting import print_report
+from validation_rules import QualityValidator
 
 
-class QualityLevel(Enum):
-    """Quality levels for validation"""
-    CRITICAL = "critical"
-    HIGH = "high"
-    MEDIUM = "medium"
-    LOW = "low"
-
-
-class ValidationStatus(Enum):
-    """Validation status results"""
-    PASSED = "passed"
-    FAILED = "failed"
-    WARNING = "warning"
-    SKIPPED = "skipped"
-
-
-@dataclass
-class QualityMetric:
-    """Quality metric definition"""
-    name: str
-    description: str
-    threshold: float
-    current_value: float = 0.0
-    status: ValidationStatus = ValidationStatus.SKIPPED
-    level: QualityLevel = QualityLevel.MEDIUM
-
-
-@dataclass
-class ValidationResult:
-    """Validation result for a quality check"""
-    check_name: str
-    status: ValidationStatus
-    message: str
-    details: Dict[str, Any] = field(default_factory=dict)
-    timestamp: float = field(default_factory=time.time)
-    level: QualityLevel = QualityLevel.MEDIUM
-
-
-@dataclass
-class QualityReport:
-    """Comprehensive quality report"""
-    overall_score: float
-    total_checks: int
-    passed_checks: int
-    failed_checks: int
-    warning_checks: int
-    skipped_checks: int
-    validation_results: List[ValidationResult] = field(default_factory=list)
-    quality_metrics: List[QualityMetric] = field(default_factory=list)
-    timestamp: float = field(default_factory=time.time)
-
-
-class QualityValidator:
-    """Main quality validation class"""
-    
-    def __init__(self, project_root: Path):
-        self.project_root = project_root
-        self.src_dir = project_root / "src"
-        self.tests_dir = project_root / "tests"
-        self.validation_results: List[ValidationResult] = []
-        self.quality_metrics: List[QualityMetric] = []
-        
-        logger.info(f"Quality Validator initialized for project: {project_root}")
-        
-    def run_comprehensive_validation(self) -> QualityReport:
-        """Run comprehensive quality validation"""
-        logger.info("Starting comprehensive quality validation")
-        
-        # Run all quality checks
-        self._validate_code_quality()
-        self._validate_testing_quality()
-        self._validate_integration_quality()
-        self._validate_quality_standards()
-        
-        # Calculate overall quality score
-        overall_score = self._calculate_overall_score()
-        
-        # Generate quality report
-        report = QualityReport(
-            overall_score=overall_score,
-            total_checks=len(self.validation_results),
-            passed_checks=len([r for r in self.validation_results if r.status == ValidationStatus.PASSED]),
-            failed_checks=len([r for r in self.validation_results if r.status == ValidationStatus.FAILED]),
-            warning_checks=len([r for r in self.validation_results if r.status == ValidationStatus.WARNING]),
-            skipped_checks=len([r for r in self.validation_results if r.status == ValidationStatus.SKIPPED]),
-            validation_results=self.validation_results,
-            quality_metrics=self.quality_metrics
-        )
-        
-        logger.info(f"Quality validation completed. Overall score: {overall_score:.2f}%")
-        return report
-    
-    def _validate_code_quality(self) -> None:
-        """Validate code quality standards"""
-        logger.info("Validating code quality standards")
-        
-        # Check SRP compliance
-        self._check_srp_compliance()
-        
-        # Check OOP design
-        self._check_oop_design()
-        
-        # Check line count compliance
-        self._check_line_count_compliance()
-        
-        # Check error handling
-        self._check_error_handling()
-        
-        # Check documentation
-        self._check_documentation()
-        
-        # Check import stability
-        self._check_import_stability()
-    
-    def _validate_testing_quality(self) -> None:
-        """Validate testing quality standards"""
-        logger.info("Validating testing quality standards")
-        
-        # Check test coverage
-        self._check_test_coverage()
-        
-        # Check smoke test coverage
-        self._check_smoke_test_coverage()
-        
-        # Check CLI interface coverage
-        self._check_cli_interface_coverage()
-        
-        # Check test execution
-        self._check_test_execution()
-        
-        # Check test organization
-        self._check_test_organization()
-    
-    def _validate_integration_quality(self) -> None:
-        """Validate integration quality standards"""
-        logger.info("Validating integration quality standards")
-        
-        # Check system startup
-        self._check_system_startup()
-        
-        # Check cross-component communication
-        self._check_cross_component_communication()
-        
-        # Check data flow
-        self._check_data_flow()
-        
-        # Check error recovery
-        self._check_error_recovery()
-    
-    def _validate_quality_standards(self) -> None:
-        """Validate quality standards compliance"""
-        logger.info("Validating quality standards compliance")
-        
-        # Check quality framework
-        self._check_quality_framework()
-        
-        # Check quality monitoring
-        self._check_quality_monitoring()
-        
-        # Check quality gates
-        self._check_quality_gates()
-    
-    def _check_srp_compliance(self) -> None:
-        """Check Single Responsibility Principle compliance"""
-        try:
-            srp_violations = 0
-            total_classes = 0
-            
-            for py_file in self.src_dir.rglob("*.py"):
-                if py_file.name.startswith("__"):
-                    continue
-                    
-                try:
-                    with open(py_file, 'r', encoding='utf-8') as f:
-                        content = f.read()
-                    
-                    tree = ast.parse(content)
-                    classes = [node for node in ast.walk(tree) if isinstance(node, ast.ClassDef)]
-                    
-                    for class_node in classes:
-                        total_classes += 1
-                        methods = [node for node in class_node.body if isinstance(node, ast.FunctionDef)]
-                        
-                        # Simple SRP check: if class has too many methods, it might violate SRP
-                        if len(methods) > 15:  # Threshold for potential SRP violation
-                            srp_violations += 1
-                            
-                except (SyntaxError, UnicodeDecodeError):
-                    continue
-            
-            srp_compliance = ((total_classes - srp_violations) / total_classes * 100) if total_classes > 0 else 100
-            
-            metric = QualityMetric(
-                name="SRP Compliance",
-                description="Single Responsibility Principle compliance percentage",
-                threshold=80.0,
-                current_value=srp_compliance,
-                level=QualityLevel.CRITICAL
-            )
-            
-            if srp_compliance >= 80:
-                status = ValidationStatus.PASSED
-                message = f"SRP compliance: {srp_compliance:.1f}% (≥80%)"
-            else:
-                status = ValidationStatus.FAILED
-                message = f"SRP compliance: {srp_compliance:.1f}% (<80%) - {srp_violations} potential violations"
-            
-            self.quality_metrics.append(metric)
-            self.validation_results.append(ValidationResult(
-                check_name="SRP Compliance Check",
-                status=status,
-                message=message,
-                details={"compliance_percentage": srp_compliance, "violations": srp_violations},
-                level=QualityLevel.CRITICAL
-            ))
-            
-        except Exception as e:
-            logger.error(f"Error checking SRP compliance: {e}")
-            self.validation_results.append(ValidationResult(
-                check_name="SRP Compliance Check",
-                status=ValidationStatus.FAILED,
-                message=f"Error checking SRP compliance: {e}",
-                level=QualityLevel.CRITICAL
-            ))
-    
-    def _check_oop_design(self) -> None:
-        """Check Object-Oriented Design compliance"""
-        try:
-            oop_violations = 0
-            total_files = 0
-            
-            for py_file in self.src_dir.rglob("*.py"):
-                if py_file.name.startswith("__"):
-                    continue
-                    
-                total_files += 1
-                
-                try:
-                    with open(py_file, 'r', encoding='utf-8') as f:
-                        content = f.read()
-                    
-                    # Check if file has classes
-                    if "class " not in content and "def " in content:
-                        # File has functions but no classes - potential OOP violation
-                        oop_violations += 1
-                        
-                except (UnicodeDecodeError):
-                    continue
-            
-            oop_compliance = ((total_files - oop_violations) / total_files * 100) if total_files > 0 else 100
-            
-            metric = QualityMetric(
-                name="OOP Design Compliance",
-                description="Object-Oriented Design compliance percentage",
-                threshold=90.0,
-                current_value=oop_compliance,
-                level=QualityLevel.HIGH
-            )
-            
-            if oop_compliance >= 90:
-                status = ValidationStatus.PASSED
-                message = f"OOP design compliance: {oop_compliance:.1f}% (≥90%)"
-            else:
-                status = ValidationStatus.FAILED
-                message = f"OOP design compliance: {oop_compliance:.1f}% (<90%) - {oop_violations} violations"
-            
-            self.quality_metrics.append(metric)
-            self.validation_results.append(ValidationResult(
-                check_name="OOP Design Check",
-                status=status,
-                message=message,
-                details={"compliance_percentage": oop_compliance, "violations": oop_violations},
-                level=QualityLevel.HIGH
-            ))
-            
-        except Exception as e:
-            logger.error(f"Error checking OOP design: {e}")
-            self.validation_results.append(ValidationResult(
-                check_name="OOP Design Check",
-                status=ValidationStatus.FAILED,
-                message=f"Error checking OOP design: {e}",
-                level=QualityLevel.HIGH
-            ))
-    
-    def _check_line_count_compliance(self) -> None:
-        """Check line count compliance"""
-        try:
-            violations = 0
-            total_files = 0
-            
-            for py_file in self.src_dir.rglob("*.py"):
-                if py_file.name.startswith("__"):
-                    continue
-                    
-                total_files += 1
-                
-                try:
-                    with open(py_file, 'r', encoding='utf-8') as f:
-                        lines = f.readlines()
-                    
-                    # Count non-empty lines
-                    code_lines = len([line for line in lines if line.strip() and not line.strip().startswith('#')])
-                    
-                    # Check if file exceeds 400 LOC limit
-                    if code_lines > 400:
-                        violations += 1
-                        
-                except (UnicodeDecodeError):
-                    continue
-            
-            compliance = ((total_files - violations) / total_files * 100) if total_files > 0 else 100
-            
-            metric = QualityMetric(
-                name="Line Count Compliance",
-                description="Line count compliance percentage (≤400 LOC)",
-                threshold=80.0,
-                current_value=compliance,
-                level=QualityLevel.MEDIUM
-            )
-            
-            if compliance >= 80:
-                status = ValidationStatus.PASSED
-                message = f"Line count compliance: {compliance:.1f}% (≥80%)"
-            else:
-                status = ValidationStatus.WARNING
-                message = f"Line count compliance: {compliance:.1f}% (<80%) - {violations} files exceed 400 LOC"
-            
-            self.quality_metrics.append(metric)
-            self.validation_results.append(ValidationResult(
-                check_name="Line Count Compliance Check",
-                status=status,
-                message=message,
-                details={"compliance_percentage": compliance, "violations": violations},
-                level=QualityLevel.MEDIUM
-            ))
-            
-        except Exception as e:
-            logger.error(f"Error checking line count compliance: {e}")
-            self.validation_results.append(ValidationResult(
-                check_name="Line Count Compliance Check",
-                status=ValidationStatus.FAILED,
-                message=f"Error checking line count compliance: {e}",
-                level=QualityLevel.MEDIUM
-            ))
-    
-    def _check_test_coverage(self) -> None:
-        """Check test coverage"""
-        try:
-            # Run pytest with coverage, excluding problematic files
-            result = subprocess.run(
-                ["python", "-m", "pytest", "--cov=src", "--cov-report=json", "--cov-fail-under=0"],
-                capture_output=True,
-                text=True,
-                cwd=self.project_root,
-                timeout=300
-            )
-            
-            if result.returncode == 0:
-                # Parse coverage report
-                try:
-                    coverage_data = json.loads(result.stdout)
-                    total_coverage = coverage_data.get("totals", {}).get("percent_covered", 0)
-                except (json.JSONDecodeError, KeyError):
-                    # Fallback: try to extract coverage from stderr or stdout
-                    output = result.stdout + result.stderr
-                    if "TOTAL" in output:
-                        # Look for coverage percentage in output
-                        coverage_match = re.search(r'TOTAL\s+\d+\s+\d+\s+(\d+)%', output)
-                        if coverage_match:
-                            total_coverage = float(coverage_match.group(1))
-                        else:
-                            total_coverage = 0
-                    else:
-                        total_coverage = 0
-                
-                metric = QualityMetric(
-                    name="Test Coverage",
-                    description="Test coverage percentage",
-                    threshold=80.0,
-                    current_value=total_coverage,
-                    level=QualityLevel.CRITICAL
-                )
-                
-                if total_coverage >= 80:
-                    status = ValidationStatus.PASSED
-                    message = f"Test coverage: {total_coverage:.1f}% (≥80%)"
-                else:
-                    status = ValidationStatus.FAILED
-                    message = f"Test coverage: {total_coverage:.1f}% (<80%) - Below threshold"
-                
-                self.quality_metrics.append(metric)
-                self.validation_results.append(ValidationResult(
-                    check_name="Test Coverage Check",
-                    status=status,
-                    message=message,
-                    details={"coverage_percentage": total_coverage},
-                    level=QualityLevel.CRITICAL
-                ))
-            else:
-                # Try alternative coverage approach
-                try:
-                    # Run a simpler coverage check
-                    simple_result = subprocess.run(
-                        ["python", "-m", "coverage", "run", "--source=src", "-m", "pytest", "--collect-only"],
-                        capture_output=True,
-                        text=True,
-                        cwd=self.project_root,
-                        timeout=120
-                    )
-                    
-                    if simple_result.returncode == 0:
-                        # Estimate coverage based on test collection
-                        test_files = list(self.tests_dir.rglob("*.py"))
-                        source_files = list(self.src_dir.rglob("*.py"))
-                        
-                        # Simple estimation: if we have tests for most source files
-                        estimated_coverage = min(75.0, (len(test_files) / max(len(source_files), 1)) * 100)
-                        
-                        metric = QualityMetric(
-                            name="Test Coverage (Estimated)",
-                            description="Estimated test coverage percentage",
-                            threshold=80.0,
-                            current_value=estimated_coverage,
-                            level=QualityLevel.CRITICAL
-                        )
-                        
-                        status = ValidationStatus.WARNING
-                        message = f"Test coverage (estimated): {estimated_coverage:.1f}% - Coverage parsing failed, using estimation"
-                        
-                        self.quality_metrics.append(metric)
-                        self.validation_results.append(ValidationResult(
-                            check_name="Test Coverage Check",
-                            status=status,
-                            message=message,
-                            details={"coverage_percentage": estimated_coverage, "method": "estimation"},
-                            level=QualityLevel.CRITICAL
-                        ))
-                    else:
-                        self.validation_results.append(ValidationResult(
-                            check_name="Test Coverage Check",
-                            status=ValidationStatus.FAILED,
-                            message=f"Test coverage check failed: {simple_result.stderr}",
-                            level=QualityLevel.CRITICAL
-                        ))
-                        
-                except Exception as e:
-                    self.validation_results.append(ValidationResult(
-                        check_name="Test Coverage Check",
-                        status=ValidationStatus.FAILED,
-                        message=f"Test coverage check failed: {e}",
-                        level=QualityLevel.CRITICAL
-                    ))
-                    
-        except Exception as e:
-            logger.error(f"Error checking test coverage: {e}")
-            self.validation_results.append(ValidationResult(
-                check_name="Test Coverage Check",
-                status=ValidationStatus.FAILED,
-                message=f"Error checking test coverage: {e}",
-                level=QualityLevel.CRITICAL
-            ))
-    
-    def _check_smoke_test_coverage(self) -> None:
-        """Check smoke test coverage"""
-        try:
-            smoke_test_files = list(self.tests_dir.rglob("*smoke*.py"))
-            smoke_test_count = len(smoke_test_files)
-            
-            # Count source files that should have smoke tests
-            source_files = list(self.src_dir.rglob("*.py"))
-            source_count = len([f for f in source_files if not f.name.startswith("__")])
-            
-            # Estimate smoke test coverage (this is a simplified check)
-            estimated_coverage = (smoke_test_count / max(source_count, 1)) * 100
-            
-            metric = QualityMetric(
-                name="Smoke Test Coverage",
-                description="Smoke test coverage percentage",
-                threshold=50.0,  # Lower threshold for smoke tests
-                current_value=estimated_coverage,
-                level=QualityLevel.HIGH
-            )
-            
-            if estimated_coverage >= 50:
-                status = ValidationStatus.PASSED
-                message = f"Smoke test coverage: {estimated_coverage:.1f}% (≥50%) - {smoke_test_count} smoke test files"
-            else:
-                status = ValidationStatus.WARNING
-                message = f"Smoke test coverage: {estimated_coverage:.1f}% (<50%) - {smoke_test_count} smoke test files"
-            
-            self.quality_metrics.append(metric)
-            self.validation_results.append(ValidationResult(
-                check_name="Smoke Test Coverage Check",
-                status=status,
-                message=message,
-                details={"coverage_percentage": estimated_coverage, "smoke_test_files": smoke_test_count},
-                level=QualityLevel.HIGH
-            ))
-            
-        except Exception as e:
-            logger.error(f"Error checking smoke test coverage: {e}")
-            self.validation_results.append(ValidationResult(
-                check_name="Smoke Test Coverage Check",
-                status=ValidationStatus.FAILED,
-                message=f"Error checking smoke test coverage: {e}",
-                level=QualityLevel.HIGH
-            ))
-    
-    def _check_cli_interface_coverage(self) -> None:
-        """Check CLI interface coverage"""
-        try:
-            cli_files = 0
-            total_files = 0
-            
-            for py_file in self.src_dir.rglob("*.py"):
-                if py_file.name.startswith("__"):
-                    continue
-                    
-                total_files += 1
-                
-                try:
-                    with open(py_file, 'r', encoding='utf-8') as f:
-                        content = f.read()
-                    
-                    # Check if file has CLI interface (main function or argparse)
-                    if "if __name__ == '__main__'" in content or "argparse" in content or "click" in content or "typer" in content:
-                        cli_files += 1
-                        
-                except (UnicodeDecodeError):
-                    continue
-            
-            cli_coverage = (cli_files / total_files * 100) if total_files > 0 else 0
-            
-            metric = QualityMetric(
-                name="CLI Interface Coverage",
-                description="CLI interface coverage percentage",
-                threshold=70.0,
-                current_value=cli_coverage,
-                level=QualityLevel.HIGH
-            )
-            
-            if cli_coverage >= 70:
-                status = ValidationStatus.PASSED
-                message = f"CLI interface coverage: {cli_coverage:.1f}% (≥70%) - {cli_files} files with CLI"
-            else:
-                status = ValidationStatus.WARNING
-                message = f"CLI interface coverage: {cli_coverage:.1f}% (<70%) - {cli_files} files with CLI"
-            
-            self.quality_metrics.append(metric)
-            self.validation_results.append(ValidationResult(
-                check_name="CLI Interface Coverage Check",
-                status=status,
-                message=message,
-                details={"coverage_percentage": cli_coverage, "cli_files": cli_files},
-                level=QualityLevel.HIGH
-            ))
-            
-        except Exception as e:
-            logger.error(f"Error checking CLI interface coverage: {e}")
-            self.validation_results.append(ValidationResult(
-                check_name="CLI Interface Coverage Check",
-                status=ValidationStatus.FAILED,
-                message=f"Error checking CLI interface coverage: {e}",
-                level=QualityLevel.HIGH
-            ))
-    
-    def _check_test_execution(self) -> None:
-        """Check test execution"""
-        try:
-            # Run a simple test to check execution
-            result = subprocess.run(
-                ["python", "-m", "pytest", "--collect-only", "-q"],
-                capture_output=True,
-                text=True,
-                cwd=self.project_root,
-                timeout=60
-            )
-            
-            if result.returncode == 0:
-                # Check for warnings in output
-                warnings = result.stderr.count("Warning") + result.stderr.count("warning")
-                
-                if warnings == 0:
-                    status = ValidationStatus.PASSED
-                    message = "Test execution: No warnings or errors detected"
-                else:
-                    status = ValidationStatus.WARNING
-                    message = f"Test execution: {warnings} warnings detected"
-                
-                self.validation_results.append(ValidationResult(
-                    check_name="Test Execution Check",
-                    status=status,
-                    message=message,
-                    details={"warnings": warnings},
-                    level=QualityLevel.MEDIUM
-                ))
-            else:
-                # Try alternative test execution approach
-                try:
-                    # Try running a simple test discovery
-                    simple_result = subprocess.run(
-                        ["python", "-c", "import pytest; print('pytest available')"],
-                        capture_output=True,
-                        text=True,
-                        cwd=self.project_root,
-                        timeout=30
-                    )
-                    
-                    if simple_result.returncode == 0:
-                        # Check if we can at least discover tests
-                        discover_result = subprocess.run(
-                            ["python", "-m", "pytest", "--collect-only", "--tb=no"],
-                            capture_output=True,
-                            text=True,
-                            cwd=self.project_root,
-                            timeout=60
-                        )
-                        
-                        if discover_result.returncode == 0:
-                            status = ValidationStatus.WARNING
-                            message = "Test execution: Tests can be discovered but execution has issues"
-                        else:
-                            status = ValidationStatus.WARNING
-                            message = f"Test execution: Test discovery issues - {discover_result.stderr[:100]}"
-                    else:
-                        status = ValidationStatus.WARNING
-                        message = "Test execution: pytest availability check failed"
-                    
-                    self.validation_results.append(ValidationResult(
-                        check_name="Test Execution Check",
-                        status=status,
-                        message=message,
-                        details={"fallback_method": "alternative_check"},
-                        level=QualityLevel.MEDIUM
-                    ))
-                    
-                except Exception as e:
-                    self.validation_results.append(ValidationResult(
-                        check_name="Test Execution Check",
-                        status=ValidationStatus.WARNING,
-                        message=f"Test execution: Fallback check failed - {e}",
-                        level=QualityLevel.MEDIUM
-                    ))
-                
-        except Exception as e:
-            logger.error(f"Error checking test execution: {e}")
-            self.validation_results.append(ValidationResult(
-                check_name="Test Execution Check",
-                status=ValidationStatus.FAILED,
-                message=f"Error checking test execution: {e}",
-                level=QualityLevel.MEDIUM
-            ))
-    
-    def _check_test_organization(self) -> None:
-        """Check test organization"""
-        try:
-            # Check if tests directory has proper structure
-            test_categories = ["unit", "integration", "smoke", "performance", "security"]
-            organized_tests = 0
-            
-            for category in test_categories:
-                category_dir = self.tests_dir / category
-                if category_dir.exists() and any(category_dir.iterdir()):
-                    organized_tests += 1
-            
-            organization_score = (organized_tests / len(test_categories)) * 100
-            
-            metric = QualityMetric(
-                name="Test Organization",
-                description="Test organization score",
-                threshold=60.0,
-                current_value=organization_score,
-                level=QualityLevel.MEDIUM
-            )
-            
-            if organization_score >= 60:
-                status = ValidationStatus.PASSED
-                message = f"Test organization: {organization_score:.1f}% (≥60%) - {organized_tests}/{len(test_categories)} categories organized"
-            else:
-                status = ValidationStatus.WARNING
-                message = f"Test organization: {organization_score:.1f}% (<60%) - {organized_tests}/{len(test_categories)} categories organized"
-            
-            self.quality_metrics.append(metric)
-            self.validation_results.append(ValidationResult(
-                check_name="Test Organization Check",
-                status=status,
-                message=message,
-                details={"organization_score": organization_score, "organized_categories": organized_tests},
-                level=QualityLevel.MEDIUM
-            ))
-            
-        except Exception as e:
-            logger.error(f"Error checking test organization: {e}")
-            self.validation_results.append(ValidationResult(
-                check_name="Test Organization Check",
-                status=ValidationStatus.FAILED,
-                message=f"Error checking test organization: {e}",
-                level=QualityLevel.MEDIUM
-            ))
-    
-    def _check_error_handling(self) -> None:
-        """Check error handling"""
-        try:
-            error_handling_files = 0
-            total_files = 0
-            
-            for py_file in self.src_dir.rglob("*.py"):
-                if py_file.name.startswith("__"):
-                    continue
-                    
-                total_files += 1
-                
-                try:
-                    with open(py_file, 'r', encoding='utf-8') as f:
-                        content = f.read()
-                    
-                    # Check for error handling patterns
-                    has_error_handling = (
-                        "try:" in content and "except:" in content or
-                        "try:" in content and "finally:" in content or
-                        "raise " in content or
-                        "logging.error" in content or
-                        "logger.error" in content
-                    )
-                    
-                    if has_error_handling:
-                        error_handling_files += 1
-                        
-                except (UnicodeDecodeError):
-                    continue
-            
-            error_handling_coverage = (error_handling_files / total_files * 100) if total_files > 0 else 0
-            
-            metric = QualityMetric(
-                name="Error Handling Coverage",
-                description="Error handling coverage percentage",
-                threshold=80.0,
-                current_value=error_handling_coverage,
-                level=QualityLevel.MEDIUM
-            )
-            
-            if error_handling_coverage >= 80:
-                status = ValidationStatus.PASSED
-                message = f"Error handling coverage: {error_handling_coverage:.1f}% (≥80%) - {error_handling_files} files with error handling"
-            else:
-                status = ValidationStatus.WARNING
-                message = f"Error handling coverage: {error_handling_coverage:.1f}% (<80%) - {error_handling_files} files with error handling"
-            
-            self.quality_metrics.append(metric)
-            self.validation_results.append(ValidationResult(
-                check_name="Error Handling Check",
-                status=status,
-                message=message,
-                details={"coverage_percentage": error_handling_coverage, "files_with_error_handling": error_handling_files},
-                level=QualityLevel.MEDIUM
-            ))
-            
-        except Exception as e:
-            logger.error(f"Error checking error handling: {e}")
-            self.validation_results.append(ValidationResult(
-                check_name="Error Handling Check",
-                status=ValidationStatus.FAILED,
-                message=f"Error checking error handling: {e}",
-                level=QualityLevel.MEDIUM
-            ))
-    
-    def _check_documentation(self) -> None:
-        """Check documentation"""
-        try:
-            documented_files = 0
-            total_files = 0
-            
-            for py_file in self.src_dir.rglob("*.py"):
-                if py_file.name.startswith("__"):
-                    continue
-                    
-                total_files += 1
-                
-                try:
-                    with open(py_file, 'r', encoding='utf-8') as f:
-                        content = f.read()
-                    
-                    # Check for documentation patterns
-                    has_documentation = (
-                        '"""' in content or "'''" in content or  # Docstrings
-                        "# " in content or  # Comments
-                        "def " in content and ":" in content and any(line.strip().startswith("#") for line in content.split('\n'))  # Functions with comments
-                    )
-                    
-                    if has_documentation:
-                        documented_files += 1
-                        
-                except (UnicodeDecodeError):
-                    continue
-            
-            documentation_coverage = (documented_files / total_files * 100) if total_files > 0 else 0
-            
-            metric = QualityMetric(
-                name="Documentation Coverage",
-                description="Documentation coverage percentage",
-                threshold=70.0,
-                current_value=documentation_coverage,
-                level=QualityLevel.MEDIUM
-            )
-            
-            if documentation_coverage >= 70:
-                status = ValidationStatus.PASSED
-                message = f"Documentation coverage: {documentation_coverage:.1f}% (≥70%) - {documented_files} files documented"
-            else:
-                status = ValidationStatus.WARNING
-                message = f"Documentation coverage: {documentation_coverage:.1f}% (<70%) - {documented_files} files documented"
-            
-            self.quality_metrics.append(metric)
-            self.validation_results.append(ValidationResult(
-                check_name="Documentation Check",
-                status=status,
-                message=message,
-                details={"coverage_percentage": documentation_coverage, "documented_files": documented_files},
-                level=QualityLevel.MEDIUM
-            ))
-            
-        except Exception as e:
-            logger.error(f"Error checking documentation: {e}")
-            self.validation_results.append(ValidationResult(
-                check_name="Documentation Check",
-                status=ValidationStatus.FAILED,
-                message=f"Error checking documentation: {e}",
-                level=QualityLevel.MEDIUM
-            ))
-    
-    def _check_import_stability(self) -> None:
-        """Check import stability"""
-        try:
-            import_errors = 0
-            total_files = 0
-            
-            for py_file in self.src_dir.rglob("*.py"):
-                if py_file.name.startswith("__"):
-                    continue
-                    
-                total_files += 1
-                
-                try:
-                    with open(py_file, 'r', encoding='utf-8') as f:
-                        content = f.read()
-                    
-                    # Try to parse imports to check for syntax errors
-                    tree = ast.parse(content)
-                    imports = [node for node in ast.walk(tree) if isinstance(node, (ast.Import, ast.ImportFrom))]
-                    
-                    # Check for potential import issues
-                    for imp in imports:
-                        if isinstance(imp, ast.Import):
-                            for alias in imp.names:
-                                if alias.name.startswith('_') and not alias.name.startswith('__'):
-                                    import_errors += 1  # Potential private import
-                        elif isinstance(imp, ast.ImportFrom):
-                            if imp.module and imp.module.startswith('_') and not imp.module.startswith('__'):
-                                import_errors += 1  # Potential private module import
-                        
-                except (SyntaxError, UnicodeDecodeError):
-                    import_errors += 1  # Syntax error indicates import problems
-                    continue
-            
-            import_stability = ((total_files - import_errors) / total_files * 100) if total_files > 0 else 100
-            
-            metric = QualityMetric(
-                name="Import Stability",
-                description="Import stability percentage",
-                threshold=90.0,
-                current_value=import_stability,
-                level=QualityLevel.MEDIUM
-            )
-            
-            if import_stability >= 90:
-                status = ValidationStatus.PASSED
-                message = f"Import stability: {import_stability:.1f}% (≥90%) - {import_errors} potential issues"
-            else:
-                status = ValidationStatus.WARNING
-                message = f"Import stability: {import_stability:.1f}% (<90%) - {import_errors} potential issues"
-            
-            self.quality_metrics.append(metric)
-            self.validation_results.append(ValidationResult(
-                check_name="Import Stability Check",
-                status=status,
-                message=message,
-                details={"stability_percentage": import_stability, "potential_issues": import_errors},
-                level=QualityLevel.MEDIUM
-            ))
-            
-        except Exception as e:
-            logger.error(f"Error checking import stability: {e}")
-            self.validation_results.append(ValidationResult(
-                check_name="Import Stability Check",
-                status=ValidationStatus.FAILED,
-                message=f"Error checking import stability: {e}",
-                level=QualityLevel.MEDIUM
-            ))
-    
-    def _check_system_startup(self) -> None:
-        """Check system startup"""
-        try:
-            # Check if key system files can be imported without errors
-            startup_files = [
-                "src/core/__init__.py",
-                "src/services/__init__.py",
-                "src/testing/__init__.py"
-            ]
-            
-            startup_success = 0
-            for file_path in startup_files:
-                try:
-                    if (self.project_root / file_path).exists():
-                        # Try to import the module
-                        module_path = str(file_path).replace('/', '.').replace('\\', '.')[:-3]
-                        __import__(module_path)
-                        startup_success += 1
-                except Exception:
-                    continue
-            
-            startup_score = (startup_success / len(startup_files)) * 100
-            
-            metric = QualityMetric(
-                name="System Startup",
-                description="System startup success rate",
-                threshold=80.0,
-                current_value=startup_score,
-                level=QualityLevel.MEDIUM
-            )
-            
-            if startup_score >= 80:
-                status = ValidationStatus.PASSED
-                message = f"System startup: {startup_score:.1f}% (≥80%) - {startup_success}/{len(startup_files)} modules start successfully"
-            else:
-                status = ValidationStatus.WARNING
-                message = f"System startup: {startup_score:.1f}% (<80%) - {startup_success}/{len(startup_files)} modules start successfully"
-            
-            self.quality_metrics.append(metric)
-            self.validation_results.append(ValidationResult(
-                check_name="System Startup Check",
-                status=status,
-                message=message,
-                details={"startup_score": startup_score, "successful_modules": startup_success},
-                level=QualityLevel.MEDIUM
-            ))
-            
-        except Exception as e:
-            logger.error(f"Error checking system startup: {e}")
-            self.validation_results.append(ValidationResult(
-                check_name="System Startup Check",
-                status=ValidationStatus.FAILED,
-                message=f"Error checking system startup: {e}",
-                level=QualityLevel.MEDIUM
-            ))
-    
-    def _check_cross_component_communication(self) -> None:
-        """Check cross-component communication"""
-        try:
-            # Check for communication patterns between components
-            communication_files = 0
-            total_files = 0
-            
-            for py_file in self.src_dir.rglob("*.py"):
-                if py_file.name.startswith("__"):
-                    continue
-                    
-                total_files += 1
-                
-                try:
-                    with open(py_file, 'r', encoding='utf-8') as f:
-                        content = f.read()
-                    
-                    # Check for cross-component communication patterns
-                    has_communication = (
-                        "import " in content or
-                        "from " in content or
-                        "manager" in content or
-                        "service" in content or
-                        "client" in content or
-                        "api" in content.lower()
-                    )
-                    
-                    if has_communication:
-                        communication_files += 1
-                        
-                except (UnicodeDecodeError):
-                    continue
-            
-            communication_coverage = (communication_files / total_files * 100) if total_files > 0 else 0
-            
-            metric = QualityMetric(
-                name="Cross-Component Communication",
-                description="Cross-component communication coverage",
-                threshold=60.0,
-                current_value=communication_coverage,
-                level=QualityLevel.MEDIUM
-            )
-            
-            if communication_coverage >= 60:
-                status = ValidationStatus.PASSED
-                message = f"Cross-component communication: {communication_coverage:.1f}% (≥60%) - {communication_files} files with communication patterns"
-            else:
-                status = ValidationStatus.WARNING
-                message = f"Cross-component communication: {communication_coverage:.1f}% (<60%) - {communication_files} files with communication patterns"
-            
-            self.quality_metrics.append(metric)
-            self.validation_results.append(ValidationResult(
-                check_name="Cross-Component Communication Check",
-                status=status,
-                message=message,
-                details={"coverage_percentage": communication_coverage, "communication_files": communication_files},
-                level=QualityLevel.MEDIUM
-            ))
-            
-        except Exception as e:
-            logger.error(f"Error checking cross-component communication: {e}")
-            self.validation_results.append(ValidationResult(
-                check_name="Cross-Component Communication Check",
-                status=ValidationStatus.FAILED,
-                message=f"Error checking cross-component communication: {e}",
-                level=QualityLevel.MEDIUM
-            ))
-    
-    def _check_data_flow(self) -> None:
-        """Check data flow"""
-        try:
-            # Check for data flow patterns
-            data_flow_files = 0
-            total_files = 0
-            
-            for py_file in self.src_dir.rglob("*.py"):
-                if py_file.name.startswith("__"):
-                    continue
-                    
-                total_files += 1
-                
-                try:
-                    with open(py_file, 'r', encoding='utf-8') as f:
-                        content = f.read()
-                    
-                    # Check for data flow patterns
-                    has_data_flow = (
-                        "def " in content and "return " in content or
-                        "class " in content and "def " in content or
-                        "=" in content and "(" in content or
-                        "[" in content and "]" in content or
-                        "{" in content and "}" in content
-                    )
-                    
-                    if has_data_flow:
-                        data_flow_files += 1
-                        
-                except (UnicodeDecodeError):
-                    continue
-            
-            data_flow_coverage = (data_flow_files / total_files * 100) if total_files > 0 else 0
-            
-            metric = QualityMetric(
-                name="Data Flow",
-                description="Data flow pattern coverage",
-                threshold=80.0,
-                current_value=data_flow_coverage,
-                level=QualityLevel.MEDIUM
-            )
-            
-            if data_flow_coverage >= 80:
-                status = ValidationStatus.PASSED
-                message = f"Data flow coverage: {data_flow_coverage:.1f}% (≥80%) - {data_flow_files} files with data flow patterns"
-            else:
-                status = ValidationStatus.WARNING
-                message = f"Data flow coverage: {data_flow_coverage:.1f}% (<80%) - {data_flow_files} files with data flow patterns"
-            
-            self.quality_metrics.append(metric)
-            self.validation_results.append(ValidationResult(
-                check_name="Data Flow Check",
-                status=status,
-                message=message,
-                details={"coverage_percentage": data_flow_coverage, "data_flow_files": data_flow_files},
-                level=QualityLevel.MEDIUM
-            ))
-            
-        except Exception as e:
-            logger.error(f"Error checking data flow: {e}")
-            self.validation_results.append(ValidationResult(
-                check_name="Data Flow Check",
-                status=ValidationStatus.FAILED,
-                message=f"Error checking data flow: {e}",
-                level=QualityLevel.MEDIUM
-            ))
-    
-    def _check_error_recovery(self) -> None:
-        """Check error recovery"""
-        try:
-            # Check for error recovery patterns
-            error_recovery_files = 0
-            total_files = 0
-            
-            for py_file in self.src_dir.rglob("*.py"):
-                if py_file.name.startswith("__"):
-                    continue
-                    
-                total_files += 1
-                
-                try:
-                    with open(py_file, 'r', encoding='utf-8') as f:
-                        content = f.read()
-                    
-                    # Check for error recovery patterns
-                    has_error_recovery = (
-                        "try:" in content and "except:" in content or
-                        "finally:" in content or
-                        "logging.info" in content or
-                        "logger.info" in content or
-                        "return " in content or
-                        "continue" in content
-                    )
-                    
-                    if has_error_recovery:
-                        error_recovery_files += 1
-                        
-                except (UnicodeDecodeError):
-                    continue
-            
-            error_recovery_coverage = (error_recovery_files / total_files * 100) if total_files > 0 else 0
-            
-            metric = QualityMetric(
-                name="Error Recovery",
-                description="Error recovery pattern coverage",
-                threshold=70.0,
-                current_value=error_recovery_coverage,
-                level=QualityLevel.MEDIUM
-            )
-            
-            if error_recovery_coverage >= 70:
-                status = ValidationStatus.PASSED
-                message = f"Error recovery coverage: {error_recovery_coverage:.1f}% (≥70%) - {error_recovery_files} files with error recovery patterns"
-            else:
-                status = ValidationStatus.WARNING
-                message = f"Error recovery coverage: {error_recovery_coverage:.1f}% (<70%) - {error_recovery_files} files with error recovery patterns"
-            
-            self.quality_metrics.append(metric)
-            self.validation_results.append(ValidationResult(
-                check_name="Error Recovery Check",
-                status=status,
-                message=message,
-                details={"coverage_percentage": error_recovery_coverage, "error_recovery_files": error_recovery_files},
-                level=QualityLevel.MEDIUM
-            ))
-            
-        except Exception as e:
-            logger.error(f"Error checking error recovery: {e}")
-            self.validation_results.append(ValidationResult(
-                check_name="Error Recovery Check",
-                status=ValidationStatus.FAILED,
-                message=f"Error checking error recovery: {e}",
-                level=QualityLevel.MEDIUM
-            ))
-    
-    def _check_quality_monitoring(self) -> None:
-        """Check quality monitoring"""
-        try:
-            # Check for quality monitoring patterns
-            quality_monitoring_files = 0
-            total_files = 0
-            
-            for py_file in self.src_dir.rglob("*.py"):
-                if py_file.name.startswith("__"):
-                    continue
-                    
-                total_files += 1
-                
-                try:
-                    with open(py_file, 'r', encoding='utf-8') as f:
-                        content = f.read()
-                    
-                    # Check for quality monitoring patterns
-                    has_quality_monitoring = (
-                        "quality" in content.lower() or
-                        "monitor" in content.lower() or
-                        "validation" in content.lower() or
-                        "check" in content.lower() or
-                        "test" in content.lower() or
-                        "coverage" in content.lower()
-                    )
-                    
-                    if has_quality_monitoring:
-                        quality_monitoring_files += 1
-                        
-                except (UnicodeDecodeError):
-                    continue
-            
-            quality_monitoring_coverage = (quality_monitoring_files / total_files * 100) if total_files > 0 else 0
-            
-            metric = QualityMetric(
-                name="Quality Monitoring",
-                description="Quality monitoring pattern coverage",
-                threshold=50.0,
-                current_value=quality_monitoring_coverage,
-                level=QualityLevel.MEDIUM
-            )
-            
-            if quality_monitoring_coverage >= 50:
-                status = ValidationStatus.PASSED
-                message = f"Quality monitoring coverage: {quality_monitoring_coverage:.1f}% (≥50%) - {quality_monitoring_files} files with quality monitoring patterns"
-            else:
-                status = ValidationStatus.WARNING
-                message = f"Quality monitoring coverage: {quality_monitoring_coverage:.1f}% (<50%) - {quality_monitoring_files} files with quality monitoring patterns"
-            
-            self.quality_metrics.append(metric)
-            self.validation_results.append(ValidationResult(
-                check_name="Quality Monitoring Check",
-                status=status,
-                message=message,
-                details={"coverage_percentage": quality_monitoring_coverage, "quality_monitoring_files": quality_monitoring_files},
-                level=QualityLevel.MEDIUM
-            ))
-            
-        except Exception as e:
-            logger.error(f"Error checking quality monitoring: {e}")
-            self.validation_results.append(ValidationResult(
-                check_name="Quality Monitoring Check",
-                status=ValidationStatus.FAILED,
-                message=f"Error checking quality monitoring: {e}",
-                level=QualityLevel.MEDIUM
-            ))
-    
-    def _check_quality_gates(self) -> None:
-        """Check quality gates"""
-        try:
-            # Check for quality gate patterns
-            quality_gate_files = 0
-            total_files = 0
-            
-            for py_file in self.src_dir.rglob("*.py"):
-                if py_file.name.startswith("__"):
-                    continue
-                    
-                total_files += 1
-                
-                try:
-                    with open(py_file, 'r', encoding='utf-8') as f:
-                        content = f.read()
-                    
-                    # Check for quality gate patterns
-                    has_quality_gates = (
-                        "gate" in content.lower() or
-                        "threshold" in content.lower() or
-                        "validation" in content.lower() or
-                        "check" in content.lower() or
-                        "compliance" in content.lower() or
-                        "standard" in content.lower()
-                    )
-                    
-                    if has_quality_gates:
-                        quality_gate_files += 1
-                        
-                except (UnicodeDecodeError):
-                    continue
-            
-            quality_gate_coverage = (quality_gate_files / total_files * 100) if total_files > 0 else 0
-            
-            metric = QualityMetric(
-                name="Quality Gates",
-                description="Quality gate pattern coverage",
-                threshold=40.0,
-                current_value=quality_gate_coverage,
-                level=QualityLevel.MEDIUM
-            )
-            
-            if quality_gate_coverage >= 40:
-                status = ValidationStatus.PASSED
-                message = f"Quality gate coverage: {quality_gate_coverage:.1f}% (≥40%) - {quality_gate_files} files with quality gate patterns"
-            else:
-                status = ValidationStatus.WARNING
-                message = f"Quality gate coverage: {quality_gate_coverage:.1f}% (<40%) - {quality_gate_files} files with quality gate patterns"
-            
-            self.quality_metrics.append(metric)
-            self.validation_results.append(ValidationResult(
-                check_name="Quality Gates Check",
-                status=status,
-                message=message,
-                details={"coverage_percentage": quality_gate_coverage, "quality_gate_files": quality_gate_files},
-                level=QualityLevel.MEDIUM
-            ))
-            
-        except Exception as e:
-            logger.error(f"Error checking quality gates: {e}")
-            self.validation_results.append(ValidationResult(
-                check_name="Quality Gates Check",
-                status=ValidationStatus.FAILED,
-                message=f"Error checking quality gates: {e}",
-                level=QualityLevel.MEDIUM
-            ))
-    
-    def _check_quality_framework(self) -> None:
-        """Check quality framework"""
-        try:
-            # Check if quality framework files exist
-            quality_files = [
-                "src/services/quality/__init__.py",
-                "src/services/quality/quality_monitor.py",
-                "src/services/quality/assurance_engine.py"
-            ]
-            
-            existing_files = 0
-            for file_path in quality_files:
-                if (self.project_root / file_path).exists():
-                    existing_files += 1
-            
-            framework_score = (existing_files / len(quality_files)) * 100
-            
-            metric = QualityMetric(
-                name="Quality Framework",
-                description="Quality framework completeness score",
-                threshold=80.0,
-                current_value=framework_score,
-                level=QualityLevel.HIGH
-            )
-            
-            if framework_score >= 80:
-                status = ValidationStatus.PASSED
-                message = f"Quality framework: {framework_score:.1f}% (≥80%) - {existing_files}/{len(quality_files)} files present"
-            else:
-                status = ValidationStatus.WARNING
-                message = f"Quality framework: {framework_score:.1f}% (<80%) - {existing_files}/{len(quality_files)} files present"
-            
-            self.quality_metrics.append(metric)
-            self.validation_results.append(ValidationResult(
-                check_name="Quality Framework Check",
-                status=status,
-                message=message,
-                details={"framework_score": framework_score, "existing_files": existing_files},
-                level=QualityLevel.HIGH
-            ))
-            
-        except Exception as e:
-            logger.error(f"Error checking quality framework: {e}")
-            self.validation_results.append(ValidationResult(
-                check_name="Quality Framework Check",
-                status=ValidationStatus.FAILED,
-                message=f"Error checking quality framework: {e}",
-                level=QualityLevel.HIGH
-            ))
-    
-    def _calculate_overall_score(self) -> float:
-        """Calculate overall quality score"""
-        if not self.validation_results:
-            return 0.0
-        
-        # Weight scores by level
-        level_weights = {
-            QualityLevel.CRITICAL: 1.0,
-            QualityLevel.HIGH: 0.8,
-            QualityLevel.MEDIUM: 0.6,
-            QualityLevel.LOW: 0.4
-        }
-        
-        total_weighted_score = 0.0
-        total_weight = 0.0
-        
-        for result in self.validation_results:
-            weight = level_weights.get(result.level, 0.5)
-            total_weight += weight
-            
-            if result.status == ValidationStatus.PASSED:
-                score = 100.0
-            elif result.status == ValidationStatus.WARNING:
-                score = 70.0
-            elif result.status == ValidationStatus.SKIPPED:
-                score = 50.0
-            else:  # FAILED
-                score = 0.0
-            
-            total_weighted_score += score * weight
-        
-        return (total_weighted_score / total_weight) if total_weight > 0 else 0.0
-    
-    def print_report(self, report: QualityReport) -> None:
-        """Print quality report"""
-        print("\n" + "="*80)
-        print("🎯 QUALITY VALIDATION REPORT - Agent-7 (Quality Assurance Manager)")
-        print("="*80)
-        print(f"Overall Quality Score: {report.overall_score:.2f}%")
-        print(f"Total Checks: {report.total_checks}")
-        print(f"Passed: {report.passed_checks} | Failed: {report.failed_checks} | Warnings: {report.warning_checks} | Skipped: {report.skipped_checks}")
-        print("-"*80)
-        
-        # Group results by level
-        for level in QualityLevel:
-            level_results = [r for r in report.validation_results if r.level == level]
-            if level_results:
-                print(f"\n{level.value.upper()} LEVEL CHECKS:")
-                for result in level_results:
-                    status_icon = "✅" if result.status == ValidationStatus.PASSED else "❌" if result.status == ValidationStatus.FAILED else "⚠️" if result.status == ValidationStatus.WARNING else "⏭️"
-                    print(f"  {status_icon} {result.check_name}: {result.message}")
-        
-        print("\n" + "="*80)
-        print("Quality validation completed successfully!")
-        print("="*80)
-
-
-def main():
-    """Main function for CLI execution"""
+def main() -> None:
+    """Entry point for running the quality validation suite."""
     if len(sys.argv) > 1 and sys.argv[1] == "--help":
         print("Quality Validation Script - Agent-7 (Quality Assurance Manager)")
         print("\nUsage: python quality_validation.py [--help]")
-        print("\nThis script runs comprehensive quality validation for the V2 workspace.")
-        print("It checks code quality, testing quality, integration quality, and quality standards compliance.")
+        print(
+            "\nThis script runs comprehensive quality validation for the V2 workspace."
+        )
+        print(
+            "It checks code quality, testing quality, integration quality, and quality standards compliance."
+        )
         return
-    
-    # Get project root
-    project_root = Path.cwd()
-    
-    # Initialize quality validator
+
+    project_root = get_project_root()
     validator = QualityValidator(project_root)
-    
-    # Run comprehensive validation
     report = validator.run_comprehensive_validation()
-    
-    # Print report
-    validator.print_report(report)
-    
-    # Exit with appropriate code
+    print_report(report)
+
     if report.failed_checks > 0:
         sys.exit(1)
     elif report.warning_checks > 0:
